requirements.txt
requests
python-dotenv
fastmcp
replicate
elevenlabs
openai 